OPE-124 clarify RHIME likelihood ownership
