+++ docs/reference/openghg_inversions.models.gaussian_likelihood.rst
@@ -1,0 +1,7 @@
+openghg\_inversions.models.gaussian\_likelihood
+================================================
+
+.. automodule:: openghg_inversions.models.gaussian_likelihood
+   :members:
+   :show-inheritance:
+   :undoc-members:
