--- docs/reference/openghg_inversions.models.likelihoods.rst
@@ -1,7 +1,0 @@
-openghg\_inversions.models.likelihoods
-======================================
-
-.. automodule:: openghg_inversions.models.likelihoods
-   :members:
-   :show-inheritance:
-   :undoc-members:
