+++ docs/reference/openghg_inversions.models.rst
@@ -12,6 +12,6 @@ openghg\_inversions.models
    :maxdepth: 4
 
    openghg_inversions.models.additive_sigma
-   openghg_inversions.models.likelihoods
+   openghg_inversions.models.gaussian_likelihood
    openghg_inversions.models.pollution_event
    openghg_inversions.models.state_activity
