+++ docs/reference/openghg_inversions.rhime.likelihoods.rst
@@ -1,6 +1,9 @@
 openghg\_inversions.rhime.likelihoods
 =====================================
 
+This compatibility module preserves the unchanged ``run_hbmcmc`` import path.
+PyMC graph construction belongs to the re-exported model-owned component.
+
 .. automodule:: openghg_inversions.rhime.likelihoods
    :members:
    :show-inheritance:
