+++ docs/usage/customising_rhime.rst
@@ -70,10 +70,11 @@ The complete integration is one named argument:
 
 RHIME calls the function with explicit keyword arguments while constructing
 the PyMC model: the prepared observations, completed forward-model mean,
-pollution contribution, pollution-event baseline, selected aggregation error,
-and output dimension. The function adds ``epsilon`` and the canonical observed
-variable ``y`` to the active model and returns ``y``. There is no framework
-context or likelihood-result record to construct.
+reported and minimum errors, selected aggregation error, and output dimension.
+Pollution-event-only terms remain inside the built-in pollution-event
+component. The function adds ``epsilon`` and the canonical observed variable
+``y`` to the active model and returns ``y``. There is no framework context or
+likelihood-result record to construct.
 
 Options owned only by a custom likelihood can be supplied separately with
 ``likelihood_kwargs``. RHIME expands that mapping into the callable without
@@ -112,9 +113,11 @@ The example rejects dense and low-rank aggregation covariance because it uses
 an independent Student-t distribution. Supporting those aggregation-error
 modes would require a multivariate likelihood.
 
-The installed ``rhime.likelihoods.additive_sigma_likelihood_builder`` is a
-drop-in ordinary likelihood builder. It derives sigma alignment from the
-labelled observation ``site`` and ``time`` coordinates. Optional
+The model-owned
+``models.additive_sigma.additive_sigma_likelihood_builder`` is a drop-in
+ordinary likelihood builder. It derives sigma alignment from the labelled
+observation ``site`` and ``time`` coordinates and constructs the PyMC graph in
+the same module as the additive-sigma equations. Optional
 ``sigma_prior``, ``sigma_freq``, ``sigma_per_site``, ``sigma_freq_anchor``, and
 ``no_model_error`` settings belong to that component and can be supplied in
 ``likelihood_kwargs``. This model adds absolute ``sigma**2`` directly to the
