+++ newsfragments/OPE-124.misc.md
@@ -1,0 +1,1 @@
+Move RHIME likelihood graph construction under model-owned modules and keep pollution-event-only inputs out of general likelihood builders.
