+++ openghg_inversions/models/additive_sigma.py
@@ -1,4 +1,4 @@
-r"""Observation error with an additive model-data-mismatch scale.
+r"""Model-owned likelihood with an additive model-data-mismatch scale.
 
 For reported observation-error variance :math:`s_y^2`, optional fixed mismatch
 :math:`s_{fixed}`, inferred mismatch scale :math:`\sigma`, and fixed
@@ -11,11 +11,18 @@ component constructs the independent variance
 
 An optional ``min_error`` applies the same floor on total marginal standard
 deviation as the historical likelihood.
+``add_additive_sigma_gaussian_likelihood`` owns the complete observation model
+for an explicitly supplied modelled concentration. The
+``additive_sigma_likelihood_builder`` entry point supplies the same complete
+PyMC construction through RHIME's ordinary likelihood-builder seam. Both
+public functions require an active model context and therefore live in this
+model module rather than under ``openghg_inversions.rhime``.
 """
 
 from __future__ import annotations
 
 from collections.abc import Mapping
+from numbers import Real
 from typing import Any, cast
 
 import numpy as np
@@ -24,8 +31,9 @@ import pytensor.tensor as pt
 import xarray as xr
 from pytensor.tensor.variable import TensorVariable
 
+from openghg_inversions.inversion_inputs import DatetimeLike, make_site_indicator, make_site_names
 from openghg_inversions.models.components import add_model_data, add_sigma_component
-from openghg_inversions.models.likelihoods import (
+from openghg_inversions.models.gaussian_likelihood import (
     add_aggregation_error_data,
     add_gaussian_observation_likelihood,
 )
@@ -150,8 +158,117 @@ def add_additive_sigma_gaussian_likelihood(
     )
 
 
+def _resolve_site_sigma_prior(
+    sigma_prior: Mapping[str, Any],
+    site: xr.DataArray,
+    *,
+    per_site: bool,
+) -> dict[str, Any]:
+    """Align an optional site-keyed ``sigma`` parameter to latent site order."""
+    resolved = dict(sigma_prior)
+    sigma = resolved.get("sigma")
+    if not isinstance(sigma, Mapping):
+        return resolved
+    if not per_site:
+        raise ValueError("A site-keyed sigma prior requires `sigma_per_site=True`.")
+
+    site_names = [str(value) for value in make_site_names(site).values]
+    missing = [name for name in site_names if name not in sigma]
+    if missing:
+        raise ValueError(f"Site-keyed sigma prior is missing retained site(s): {missing!r}.")
+
+    values = [sigma[name] for name in site_names]
+    if any(isinstance(value, bool) or not isinstance(value, Real) for value in values):
+        raise ValueError("Site-keyed sigma prior values must be finite positive numbers.")
+    scales = np.asarray(values, dtype=float)
+    if np.any(~np.isfinite(scales)) or np.any(scales <= 0):
+        raise ValueError("Site-keyed sigma prior values must be finite positive numbers.")
+    resolved["sigma"] = scales[:, None]
+    return resolved
+
+
+def additive_sigma_likelihood_builder(
+    *,
+    observations: xr.DataArray,
+    observation_error: xr.DataArray,
+    minimum_error: xr.DataArray,
+    aggregation_error: AggregationError,
+    mean: TensorVariable,
+    output_dim: str,
+    sigma_prior: Mapping[str, Any] | None = None,
+    sigma_freq: str | None = None,
+    sigma_per_site: bool = True,
+    sigma_freq_anchor: DatetimeLike | None = None,
+    no_model_error: bool = False,
+) -> TensorVariable:
+    """Build the additive-sigma Gaussian through RHIME's likelihood seam.
+
+    Additive mismatch variance is independent of pollution-event terms. The
+    builder derives sigma alignment from labelled observation ``site`` and
+    ``time`` coordinates, then constructs the complete likelihood in the
+    active PyMC model.
+
+    Args:
+        observations: Observed mole fractions.
+        observation_error: Reported observation-error standard deviations.
+        minimum_error: Minimum total-error standard deviations.
+        aggregation_error: Validated fixed aggregation-error representation.
+        mean: Completed forward-model concentration.
+        output_dim: Observation dimension used for named PyMC variables.
+        sigma_prior: Optional prior arguments used to construct ``sigma``. The
+            distribution's ``sigma`` parameter may map site names to scales
+            when ``sigma_per_site`` is true.
+        sigma_freq: Optional frequency for sigma periods.
+        sigma_per_site: Whether sigma varies independently by site.
+        sigma_freq_anchor: Optional anchor for fixed-duration sigma periods.
+        no_model_error: Whether to omit inferred mismatch error.
+
+    Returns:
+        The observed Gaussian variable, named ``y``. The component also adds
+        the canonical marginal error scale ``epsilon``.
+
+    Raises:
+        ValueError: If observation labels, sigma settings, or error inputs are
+            invalid.
+    """
+    site = None
+    sigma_alignment = None
+    if not no_model_error:
+        site = observations.coords.get("site")
+        if site is None or site.dims != (output_dim,):
+            raise ValueError(
+                "The additive-sigma likelihood requires an observation-aligned "
+                "'site' coordinate when model error is enabled."
+            )
+        sigma_alignment = SigmaAlignment.from_frequency(
+            make_site_indicator(site),
+            frequency=sigma_freq,
+            per_site=sigma_per_site,
+            anchor_time=sigma_freq_anchor,
+        )
+    resolved_sigma_prior = None
+    if not no_model_error:
+        assert site is not None
+        resolved_sigma_prior = _resolve_site_sigma_prior(
+            DEFAULT_ADDITIVE_SIGMA_PRIOR if sigma_prior is None else sigma_prior,
+            site,
+            per_site=sigma_per_site,
+        )
+    return add_additive_sigma_gaussian_likelihood(
+        observations=observations,
+        observation_error=observation_error,
+        minimum_error=minimum_error,
+        aggregation_error=aggregation_error,
+        mean=mean,
+        sigma_alignment=sigma_alignment,
+        sigma_prior=resolved_sigma_prior,
+        output_dim=output_dim,
+    )
+
+
 __all__ = [
     "DEFAULT_ADDITIVE_SIGMA_PRIOR",
     "FIXED_MODEL_MISMATCH",
     "add_additive_sigma_gaussian_likelihood",
+    "additive_sigma_likelihood_builder",
 ]
