+++ openghg_inversions/models/gaussian_likelihood.py
@@ -1,4 +1,11 @@
-"""Observation likelihoods for diagonal and structured covariance."""
+"""Model-owned Gaussian observation-distribution construction.
+
+This module owns the PyMC graph mechanics shared by concrete Gaussian
+likelihood components. Error models such as additive sigma and pollution-event
+scaling prepare their independent variance and fixed aggregation covariance,
+then call :func:`add_gaussian_observation_likelihood` inside the active model.
+RHIME runner seams do not construct or adapt these PyMC variables.
+"""
 
 from __future__ import annotations
